Update - Added : away to install NexoPOS from the CLI - Added :upload user profile - Added : isolated media manager

+++ database/migrations/create-tables/2021_02_23_004748_create_new_instalments_permissions.php
@@ -61,17 +61,19 @@ class CreateNewInstalmentsPermissions extends Migration
      */
     public function down()
     {
-        collect([
-            'nexopos.create.orders-instalments',
-            'nexopos.update.orders-instalments',
-            'nexopos.read.orders-instalments',
-            'nexopos.delete.orders-instalments',
-        ])->each( function( $identifier ) {
-            $permission     =   Permission::where( 'namespace', $identifier 
-                )->first();
-
-            $permission->removeFromRoles();
-            $permission->delete();
-        });
+        if( Schema::hasTable( 'nexopos_permissions' ) && Schema::hasTable( 'nexopos_roles' ) ) {
+            collect([
+                'nexopos.create.orders-instalments',
+                'nexopos.update.orders-instalments',
+                'nexopos.read.orders-instalments',
+                'nexopos.delete.orders-instalments',
+            ])->each( function( $identifier ) {
+                $permission     =   Permission::where( 'namespace', $identifier 
+                    )->first();
+    
+                $permission->removeFromRoles();
+                $permission->delete();
+            });
+        }
     }
 }
